move surf functions to DataSurfaces, rename and combine functions
- move GetAddOnOverrideSurfaceList and GetMaterialNumFromSurfNum into DataSurfaces
- move GetMaterialAddOnInput into GetMaterialData
- rename function to UpdateVariableAbsorptances
